pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudsec_challenge"
version = "1.0.0"
description = "Cloud Security Simulation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.2",    
    "uvicorn>=0.27.1",
    "pydantic>=2.6.4",
    "gradio>=5.3.0",     
    "pandas",
    "requests",
    "openai",
    "python-dotenv",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.txt"]